requirements.txt
cython
chainer==6.0.0
chainercv==0.13.1
cupy-cuda90==6.0.0
h5py
imageio
imageio-ffmpeg
ipdb
jupyter
matplotlib
mpi4py
numpy
opencv-python
pandas==0.25.3
protobuf==3.7.1
pytest
pyyaml
scipy
scikit-learn
scikit-image
tqdm
xlrd
